Added dialog spinner
